Update module files and install locations
