+++ CMakeLists.txt
@@ -325,6 +325,12 @@ set(CELERITAS_LIBRARY_OUTPUT_DIRECTORY
   "${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR}")
 set(CELERITAS_RUNTIME_OUTPUT_DIRECTORY
   "${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}")
+if(CELERITAS_USE_LArSoft)
+  set(CELERITAS_INSTALL_FHICLDIR "fcl")
+  set(CELERITAS_FHICL_OUTPUT_DIRECTORY
+    "${PROJECT_BINARY_DIR}/fcl")
+endif()
+
 
 #----------------------------------------------------------------------------#
 # DEPENDENCIES
